Address thermo-nuclear review feedback on issue remediation PR.
Table-drive activity replay display, split replay tests into scenario files under 800 lines, break bridge type circular imports, and consolidate env/string helpers.
