+++ src/cursor-edit-diff.ts
@@ -1,14 +1,7 @@
 const CURSOR_EDIT_DIFF_FIELD_ORDER = ["diffString", "diff", "unifiedDiff", "patch"] as const;
 
-export type CursorEditDiffSource = Record<string, unknown> | {
-	diffString?: string;
-	diff?: string;
-	unifiedDiff?: string;
-	patch?: string;
-};
-
-export function resolveCursorEditDiff(source: CursorEditDiffSource | undefined): string | undefined {
-	if (!source) return undefined;
+export function resolveCursorEditDiff(source: unknown): string | undefined {
+	if (!source || typeof source !== "object") return undefined;
 	const record = source as Record<string, unknown>;
 	for (const key of CURSOR_EDIT_DIFF_FIELD_ORDER) {
 		const value = record[key];
